instancing update move to subModel (#6)
