+++ native/tools/tojs/scene.ini
@@ -70,7 +70,7 @@ classes = Light DirectionalLight SpotLight SphereLight Model SubModel Pass Rende
 # functions from all classes.
 skip = Pass::[getBlocks initPassFromTarget],
        AABB::[getBoundary aabbAabb aabbFrustum aabbPlan merge transform transformExtentM4 setCenter getCenter isValid setValid setHalfExtents getHalfExtents set fromPoints],
-       SubModel::[getInstancedAttributeBlock (g|s)etInstancedWorldMatrixIndex],
+       SubModel::[getInstancedAttributeBlock (g|s)etInstancedWorldMatrixIndex (g|s)etInstancedAttributeIndex updateInstancedAttributes updateInstancedWorldMatrix],
        Model::[getEventProcessor getOctreeNode setOctreeNode updateOctree],
        SkinningModel::[uploadJointData],
        Frustum::[update type planes],
